project/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic[email]>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pillow>=9.0.0
opencv-python>=4.5.3
numpy>=1.21.2
alembic>=1.7.3